+++ AppCenter/AppCenter/AppCenter.h
@@ -3,6 +3,7 @@
 
 #import <Foundation/Foundation.h>
 
+#if __has_include(<AppCenter/MSACAppCenter.h>)
 #import <AppCenter/MSACAbstractLog.h>
 #import <AppCenter/MSACAppCenter.h>
 #import <AppCenter/MSACAppCenterErrors.h>
@@ -20,3 +21,22 @@
 #import <AppCenter/MSACServiceAbstract.h>
 #import <AppCenter/MSACWrapperLogger.h>
 #import <AppCenter/MSACWrapperSdk.h>
+#else
+#import "MSACAbstractLog.h"
+#import "MSACAppCenter.h"
+#import "MSACAppCenterErrors.h"
+#import "MSACChannelGroupProtocol.h"
+#import "MSACChannelProtocol.h"
+#import "MSACConstants+Flags.h"
+#import "MSACConstants.h"
+#import "MSACCustomProperties.h"
+#import "MSACDevice.h"
+#import "MSACEnable.h"
+#import "MSACLog.h"
+#import "MSACLogWithProperties.h"
+#import "MSACLogger.h"
+#import "MSACService.h"
+#import "MSACServiceAbstract.h"
+#import "MSACWrapperLogger.h"
+#import "MSACWrapperSdk.h"
+#endif