+++ AppCenter/AppCenterTests/MSACDBStorageTests.m
@@ -241,6 +241,37 @@ static const long kMSACTestStorageSizeMinimumUpperLimitInBytes = 40 * 1024;
   assertThatLong(counter, equalToLong(0));
 }
 
+- (void)testInitWithZeroPageSize {
+    id dbStorageMock = OCMClassMock([MSACDBStorage class]);
+    OCMStub([dbStorageMock getPageSizeInOpenedDatabase: [OCMArg anyPointer]]).andReturn(0);
+    self.sut = [[MSACDBStorage alloc] initWithSchema:self.schema version:1 filename:kMSACTestDBFileName];
+    int result = [self.sut executeQueryUsingBlock:^int(void *_Nonnull __unused db) {
+      return SQLITE_OK;
+    }];
+    XCTAssertEqual(SQLITE_ERROR, result);
+    [dbStorageMock stopMocking];
+}
+
+- (void)testMaxStorageSizeWithZeroPageSize {
+    id dbStorageMock = OCMClassMock([MSACDBStorage class]);
+    OCMStub([dbStorageMock getPageSizeInOpenedDatabase: [OCMArg anyPointer]]).andReturn(0);
+    self.sut = [[MSACDBStorage alloc] initWithSchema:self.schema version:1 filename:kMSACTestDBFileName];
+    XCTestExpectation *expectation = [self expectationWithDescription:@"Completion handler invoked."];
+    long maxCapacityInBytes = kMSACTestStorageSizeMinimumUpperLimitInBytes + 4 * 1024;
+    [self.sut setMaxStorageSize:maxCapacityInBytes
+                           completionHandler:^(BOOL success){
+                             XCTAssertFalse(success);
+                             [expectation fulfill];
+                           }];
+    [self waitForExpectationsWithTimeout:1
+                                 handler:^(NSError *_Nullable error) {
+                                   if (error) {
+                                     XCTFail(@"Expectation Failed with error: %@", error);
+                                   }
+                                 }];
+    [dbStorageMock stopMocking];
+}
+
 - (void)testGetPageSizeInOpenedDatabaseReturnsZeroWhenQueryFails {
 
   // If